# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Duet API
  version: 0.1.0
  description: |
    duetd 对外暴露的全部接口。Tauri 壳与浏览器走的是同一套 —— 壳不得通过 IPC 绕过 HTTP。

    **错误文案**：`Problem.type` 是机器可读错误码，前端据此查 i18n 词条；
    `detail` 只给开发者看，界面不展示。后端绝不返回用户可见的中文文案（见 docs/rules/i18n.md §3）。
servers:
  - url: http://127.0.0.1:{port}/v1
    variables:
      port:
        default: '7777'
security:
  - bearerAuth: []
...
paths:
...
  /works:
    get:
      tags: [work]
      operationId: listWorks
      summary: 全部工作
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [works]
                properties:
                  works:
                    type: array
                    items: { $ref: '#/components/schemas/Work' }
    post:
      tags: [work]
      operationId: startWork
      summary: 对一个项目提需求，开一个工作
      description: |
        切一个独立 worktree、建 ACP 会话、开始澄清需求。

        ★ worktree 建在 `~/.acpflows/worktrees`，**不在用户的项目目录里**
        （见 docs/plan/open-questions.md Q30）。用户把代码目录交给 Duet 时，
        并没有同意我们在他的仓库里造一堆分支和目录。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project, prompt]
              properties:
                project:
                  type: string
                  description: 项目的本地绝对路径
                prompt:
                  type: string
                  description: 用户的需求原话
                base_ref:
                  type: string
                  description: |
                    从哪儿开分支：分支名或 commit。留空时用仓库当前 HEAD。

                    ★ 让用户选基线的理由：他可能想从 `develop` 开工，
                    而当前分支上正躺着他没提交完的东西。
                  examples: [develop, 7c1de98]
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Work' }
        default: { $ref: '#/components/responses/Problem' }
...
  /works/{id}/units/{unitId}/accept:
    post:
      tags: [work]
      operationId: acceptUnit
      summary: 验收通过——提交改动并落检查点
      description: |
        ★★ **通过由用户点，不由 AI 判断。** 没有任何路径能自动走到这里。

        ★ 一条验收标准都没有证据时返回 409 `nothing_accepted`：
        允许的话，「验收」这个动作就没有内容了——用户点通过时以为自己
        核对过什么，而实际上什么都没有。

        ★ 没有改动时返回 409 `nothing_to_commit`，**不造空提交**：
        一个「验收通过」却什么都没改的单元，说明该被质疑的是那次验收。

        提交发生在**工作自己的 worktree** 上，用户的分支一字不动。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: unitId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 已提交
          content:
            application/json:
              schema:
                type: object
                required: [commit]
                properties:
                  commit:
                    type: string
                    description: 那次提交的短 sha，检查点绑着它
                    examples: ['abc1234']
        default: { $ref: '#/components/responses/Problem' }
  /works/{id}/units/{unitId}/start:
    post:
      tags: [work]
      operationId: startUnit
      summary: 让这个单元开工
      description: |
        ★★ 契约**没冻结时拒绝**（409 `contract_not_frozen`）：没冻结就开工
        的话，AI 干到一半契约变了，而它已经照着旧的那份改了十几个文件——
        产出对不上任何一版契约。

        这一轮由**单元自己派的那个角色**跑（裁定三）。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: unitId, in: path, required: true, schema: { type: string } }
      responses:
        '202': { description: 收到了，正在做 }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        duetd 启动时生成一次性 token，写入 `~/.acpflows/runtime/session.json`（0600）。
        无 token 一律 401 —— 防止本机其他程序静默驱动 Agent 写用户的代码。
...
  responses:
    Problem:
      description: 错误（RFC 9457）
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          description: |
            ★ 机器可读错误码（snake_case），**前端据此查 i18n 词条**。
            新增取值时必须同步加进 zh-CN.json 与 en-US.json，make check-i18n 会校验。
          examples: [contract_frozen, runtime_not_installed, cancel_timeout]
        title: { type: string, description: 简短英文标题，仅供开发者 }
        status: { type: integer }
        detail:
          type: string
          description: 英文调试信息。**界面不直接展示**，见 docs/rules/i18n.md §3。
        params:
          type: object
          additionalProperties: true
          description: 结构化参数，供前端词条插值。
        instance: { type: string }

    # ── 系统 ──────────────────────────────────────────────────
...
    Work:
      type: object
      required: [id, state]
      description: |
        一次工作。`state` 是**封闭枚举**，与 backend/internal/constant 的
        WorkState 一一对应；新增取值要同时改三处：本文件 + 那个常量 +
        状态迁移表（domain/model/work.go）。
      properties:
        id: { type: string, examples: ['work-01'] }
        state:
          type: string
          description: |
            ★ `initializing_failed` 是**终态、不可恢复** —— worktree 没切成
            就没有可执行的现场（docs/adr/0006 Q1）。界面上不该给「重试」按钮。
          enum:
            - initializing
            - initializing_failed
            - clarifying
            - planning
            - ready
            - executing
            - reviewing_unit
            - waiting_user
            - paused
            - completed
            - failed
        project:
          type: string
          description: 项目的本地绝对路径
        worktree:
          type: string
          description: 这个工作的独立工作区；**在用户项目之外**
        prompt:
          type: string
          description: 用户的需求原话
        title:
          type: string
          description: |
            列表里显示的名字，**取自用户提的那句需求**（截断）。

            ★★ 左栏显示 `work-01` 的话，用户看不出那条工作是干嘛的——
            而他可能同时开着五六条。

            ★ 不用 AI 起的名字：起的名字与他说的话对不上时，
            他在列表里找不到自己那条工作。
          examples: ['用户能取消正在运行的 turn，取消后现场证据要保留']

    # ── Runtime ───────────────────────────────────────────────
...